Add switch to allow passing unhashed root password
The root password is now expected to be unhashed, and the role will take
over the duty of hashing it. This requires additionally setting
`users_root_password_salt`.

The previous behaviour can be restored by setting
`users_root_password_is_hashed: yes` (in which case playbook authors can
keep passing an already hashed password to this role;
`users_root_password_salt` is not required in that case).