fix(delete): Handle case where Adfinis group does not exist
This may typically happen when the role is run on a system for the first
time and the Adfinis group has never been generated before.

We use the same approach as for the user creation: Attempt to gather
group information and store the fact of whether the group exists or not
in a variable, then conditionally do follow-up steps depending on
whether we did get any information.